requirements.txt
chainlit==0.5.0
openai==0.27.8
python-dotenv==1.0.0
tenacity==8.2.2
